Web: Extract shared camera types to camera_types_web.h
